% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{accession2url}
\alias{accession2url}
\title{get EBI url for an RUN fastq directory}
\usage{
accession2url(run_accessions)
}
\arguments{
\item{run_accessions}{character() vector of SRR, DRR, or ERR accessions}
}
\value{
named vector with the parent directory 
containing the fastq files for each accession
}
\description{
get EBI url for an RUN fastq directory
}
\examples{
accession2url(c('SRR000237','SRR123456'))


}
