% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{ftp_dir_listing}
\alias{ftp_dir_listing}
\title{return a data.frame directory listing}
\usage{
ftp_dir_listing(url)
}
\arguments{
\item{url}{character() vector of ftp urls}
}
\description{
Given a set of ftp locations, return the 
directory listing as a data.frame.
}
\details{
Note that this function must loop over the URLs, 
so it may be slow and inefficient.
}
\examples{
ftp_dir_listing(accession2url(c('SRR000237','SRR000238', 'SRR925811')))

}
