% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refclasses.R
\docType{data}
\name{Scroller}
\alias{Scroller}
\title{Class Scroller}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Scroller
}
\description{
Scroll through large result sets
}
\details{
This class is not meant to be called by an end 
user. Instead, start with a Searcher object and 
then call \code{searcher$scroll()}.
}
\keyword{datasets}
