% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sra_browse_API}
\alias{sra_browse_API}
\title{Browse the API interactively}
\usage{
sra_browse_API(version = "1.0")
}
\arguments{
\item{version}{character(1), the API version}
}
\description{
The SRA API is a swagger 2.0 compliant API and has a simple
user interface for experimenting with the API itself. The
API could be used from any language or even command-line
utilities for accessing web resources.
}
