% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sra_experiment_search}
\alias{sra_experiment_search}
\alias{sra_run_search}
\alias{sra_study_search}
\alias{sra_sample_search}
\alias{sra_full_search}
\title{search SRA metadata}
\usage{
sra_experiment_search(q = "*", start = 0, size = 10, fields = NULL)

sra_run_search(q = "*", start = 0, size = 10, fields = NULL)

sra_study_search(q = "*", start = 0, size = 10, fields = NULL)

sra_sample_search(q = "*", start = 0, size = 10, fields = NULL)

sra_full_search(q = "*", start = 0, size = 10, fields = NULL)
}
\arguments{
\item{q}{a lucene query string}

\item{start}{integer(1) for paging results (unit is the number of records, not number of pages)}

\item{size}{integer(1) for the number of results to return per page}

\item{fields}{character() vector with fields to return. The default \code{NULL}
will return all available fields}
}
\value{
a \code{tibble} of records, with some columns potentially 
    containing a list.
}
\description{
search SRA metadata
}
\section{Functions}{
\itemize{
\item \code{sra_run_search}: search runs

\item \code{sra_study_search}: search studies

\item \code{sra_sample_search}: search studies

\item \code{sra_full_search}: search studies
}}

\examples{
sra_run_search()

}
