% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sra_get_swagger_json_url}
\alias{sra_get_swagger_json_url}
\title{URL for the swagger json API description}
\usage{
sra_get_swagger_json_url(version = "1.0")
}
\arguments{
\item{version}{character(1), the API version}
}
\value{
character(1), the URL of the \code{swagger.json} file.
}
\description{
The SRA API is a swagger 2.0 compliant API. As such, it is
described by a json file that can be used to build additional
clients in many other languages. See \url{https://swagger.io/}
for more details.
}
