% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sradb_get_client}
\alias{sradb_get_client}
\title{Initialize an SRAdb API client}
\usage{
sradb_get_client()
}
\value{
an client
}
\description{
Initialize an SRAdb API client
}
\examples{
client = sradb_get_client()
# self-describing
names(client)
# self-documenting
client[[1]]

}
